requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
opencv-python
pillow
tensorflow
jupyter
ipykernel
